+++ tools/virgil-trust-provisioner/db_converter.py
@@ -1,4 +1,4 @@
-from virgil_keymanager.__main__ import converter_main
+from virgil_trust_provisioner.__main__ import converter_main
 
 if __name__ == '__main__':
     converter_main()
