+++ tools/virgil-trust-provisioner/key_signer.py
@@ -1,4 +1,4 @@
-from virgil_keymanager.__main__ import key_signer_main
+from virgil_trust_provisioner.__main__ import key_signer_main
 
 if __name__ == '__main__':
     key_signer_main()
