+++ tools/virgil-trust-provisioner/virgil-trust-provisioner.py
@@ -1,4 +1,4 @@
-from virgil_keymanager.__main__ import main
+from virgil_trust_provisioner.__main__ import main
 
 if __name__ == '__main__':
     main()
