+++ tools/virgil-trust-provisioner/virgil_trust_provisioner/core_utils/virgil_sign_extractor.py
