+++ tools/virgil-trust-provisioner/virgil_trust_provisioner/external_utils/util/emulator/dongles-cli
