+++ tools/virgil-trust-provisioner/virgil_trust_provisioner/external_utils/util/origin/dongles-cli
