+++ tools/virgil-trust-provisioner/virgil_trust_provisioner/storage/logging_extensions/signed_handler.py
