+++ tools/virgil-trust-provisioner/virgil_trust_provisioner/storage/tinydb_storage_extensions/byte_storage.py
